+++ packages/libs/client-utils/src/core/utils/with-emitter.ts
@@ -1,56 +1,36 @@
+import { MyEventTarget } from './MyEventTarget';
 import type { IEmit } from './helpers';
-import type { Any, AnyFunc, First, IfAny, Tail } from './types';
+import { asyncLock } from './helpers';
+import type { Any, UnionToIntersection } from './types';
 
-// the default EventTarget does not throw errors when dispatching events
-class MyEventTarget {
-  private _listeners: Record<string, ((data: Any) => void)[]> = {};
+type ToNextType<T extends Array<{ event: string; data: Any }>> =
+  UnionToIntersection<
+    {
+      [K in keyof T]: T[K] extends { event: infer Tag }
+        ? (event: Tag) => Promise<T[K]['data']>
+        : never;
+    }[number]
+  >;
 
-  public addEventListener(event: string, cb: (event: Any) => void) {
-    if (this._listeners[event] === undefined) {
-      this._listeners[event] = [];
-    }
-    this._listeners[event].push(cb);
-  }
-
-  public dispatchEvent(event: string, data: Any) {
-    if (this._listeners[event] === undefined) {
-      return;
-    }
-    this._listeners[event].forEach((cb) => cb(data));
-  }
-}
-
-type GeneralEvent<T> = (event: string, cb: (data: unknown) => Any) => T;
-
-type EventListenerType<
-  Events extends Array<{ event: string; data: Any }>,
-  WrapperType = Any,
-  Acc extends AnyFunc = Any,
-> = Events extends []
-  ? Acc & GeneralEvent<WrapperType>
-  : First<Events> extends { event: infer Tag; data: infer R }
-  ? EventListenerType<
-      Tail<Events> extends Any[] ? Tail<Events> : [],
-      WrapperType,
-      IfAny<
-        Acc,
-        (event: Tag, cb: (data: R) => Any) => WrapperType,
-        Acc & ((event: Tag, cb: (data: R) => Any) => WrapperType)
-      >
-    >
-  : EventListenerType<
-      Tail<Events> extends Any[] ? Tail<Events> : [],
-      WrapperType,
-      Acc
-    >;
+type ToOnType<
+  T extends Array<{ event: string; data: Any }>,
+  WrapperType,
+> = UnionToIntersection<
+  {
+    [K in keyof T]: T[K] extends { event: infer Tag; data: infer R }
+      ? (event: Tag, cb: (data: R) => Any) => WrapperType
+      : never;
+  }[number]
+>;
 
 export interface IEmitterWrapper<
   T extends Array<{ event: string; data: Any }>,
   Extra extends Array<{ event: string; data: Any }>,
   ExecReturnType,
 > {
-  on: EventListenerType<[...Extra, ...T], this>;
+  on: ToOnType<[...T, ...Extra], this>;
   execute: () => ExecReturnType;
+  executeTo: (() => ExecReturnType) & ToNextType<[...T, ...Extra]>;
 }
 
 export type WithEmitter<
@@ -69,17 +49,57 @@ export const withEmitter: WithEmitter =
   (fn) =>
   (...args: Any[]): Any => {
     const emitter = new MyEventTarget();
-    const execute = fn(((event: string) => (data: Any) => {
-      emitter.dispatchEvent(event, data);
-      return data;
+    const allEvents: Record<string, unknown> = {};
+    const execute: (...args: any) => Promise<any> = fn(((event: string) => {
+      allEvents[event] = null;
+      return async (data: Any) => {
+        allEvents[event] = data;
+        await emitter.dispatchEvent(event, data);
+        return data;
+      };
     }) as Any);
+    let executePromise: Promise<any> | null = null;
+
+    const exec = (): Promise<any> => {
+      if (executePromise !== null) {
+        return executePromise;
+      }
+      executePromise = execute(...args);
+      return executePromise;
+    };
+    const lock = asyncLock();
     const wrapper = {
       on: (event: string, cb: (data: Any) => Any) => {
         // CustomEvent is not typed correctly
         emitter.addEventListener(event, cb);
         return wrapper;
       },
-      execute: () => execute(...args),
+      executeTo: (event?: string) => {
+        if (event === undefined) {
+          lock.open();
+          return exec();
+        }
+        if (allEvents[event] !== null) {
+          return Promise.resolve(allEvents[event]);
+        }
+        const pr = new Promise((resolve, reject) => {
+          exec().catch(reject);
+
+          const resolveAndLock = (data: any) => {
+            resolve(data);
+            emitter.removeEventListener(event, resolveAndLock);
+            lock.close();
+            return lock.waitTillOpen();
+          };
+          emitter.addEventListener(event, resolveAndLock);
+        });
+
+        lock.open();
+        return pr;
+      },
+      execute: () => {
+        return wrapper.executeTo();
+      },
     };
     return wrapper;
   };