+++ packages/libs/react-ui/src/components/Divider/Divider.stories.tsx
@@ -1,10 +1,12 @@
+import { Card } from '@components/Card';
+import { withContentWidth } from '@storyDecorators';
 import type { Meta, StoryObj } from '@storybook/react';
 import React from 'react';
 import { Divider } from './Divider';
-import { ContentClass } from './stories.css';
 
 const meta: Meta = {
   title: 'Layout/Divider',
+  decorators: [withContentWidth],
   parameters: {
     status: { type: 'stable' },
     docs: {
@@ -25,7 +27,7 @@ export const Static: Story = {
   render: () => {
     return (
       <>
-        <div className={ContentClass}>
+        <Card>
           Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
           eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad
           minim veniam, quis nostrud exercitation ullamco laboris nisi ut
@@ -33,10 +35,7 @@ export const Static: Story = {
           reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
           pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
           culpa qui officia deserunt mollit anim id est laborum.
-        </div>
-
-        <Divider />
-        <div className={ContentClass}>
+          <Divider />
           Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
           eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad
           minim veniam, quis nostrud exercitation ullamco laboris nisi ut
@@ -44,7 +43,7 @@ export const Static: Story = {
           reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla
           pariatur. Excepteur sint occaecat cupidatat non proident, sunt in
           culpa qui officia deserunt mollit anim id est laborum.
-        </div>
+        </Card>
       </>
     );
   },